refactor: detail 페이지의 apis 폴더에 domain 별 폴더 추가
