+++ src/pages/Detail/apis/file/useDeleteFile.ts
