+++ src/pages/Detail/apis/file/useUploadFiles.ts
