+++ src/pages/Detail/apis/thread/useDeleteThread.ts
